Fix -Wformat mismatches in dd.c/uudecode.c: use PRIuMAX, not %ju/%zu
tools/lint.sh warn (clang -Wall -Wextra -fsyntax-only sweep) flagged 6
-Wformat findings, all on the aarch64/Linux pass only: dd.c's summary
line ("%ju+%ju records in/out") and uudecode.c's truncated-line
diagnostic ("%zu characters").

Root cause: this tree's own <stdint.h> always defines uintmax_t as
"unsigned _Int64" (unsigned long long) and size_t as "unsigned _Addr"
(also unsigned long long), regardless of target. On an LP64 Linux
target, clang's built-in notion of what %j/%z name is the host's own
"unsigned long" -- a real type mismatch against the vararg actually
passed, not a false positive. include/inttypes.h already carries
PRIuMAX (== "llu") for exactly this, but nothing in src/ used it.

The mismatch is invisible almost everywhere else in the tree because
include/stdio.h deliberately does not attribute printf/fprintf/etc.
with __attribute__((format(printf,...))) -- so -Wformat never runs
against a plain printf() call here at all. __util_diagf() (in
src/internal/util.h) is the one printf-style wrapper that does carry
that attribute, which is why these two call sites through it were the
only ones the sweep could actually see.

Fixed by switching both to PRIuMAX (casting uudecode.c's size_t
arguments to uintmax_t first, matching the cast idiom already used for
%ju in cksum.c/du.c). Verified the two findings are gone from a rerun
of tools/lint.sh warn, and that a full clean build succeeds on both
aarch64 (clang) and x86_64-win32 (tcc).

+++ src/util/uudecode.c
@@ -53,6 +53,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <errno.h>
+#include <inttypes.h>
 #include <sys/stat.h>
 #include "util.h"
 #include "uucode.h"
@@ -85,7 +86,15 @@ static int decode_line(const char *prog, const char *line, size_t have,
 	size_t pos, written = 0;
 
 	if (have < needed) {
-		__util_diagf("%s: truncated data line (need %zu characters, got %zu)\n", prog, needed, have);
+		/* %zu assumes the host libc's own size_t; this tree's size_t is
+		 * always "unsigned _Addr" (unsigned long long) regardless of
+		 * target, which a bare 'z' length modifier does not name on an
+		 * LP64 host. __util_diagf is the one printf-family function
+		 * here with an explicit format(printf) attribute (see dd.c's
+		 * identical %ju fix and PRIuMAX's own comment in
+		 * include/inttypes.h), so cast to uintmax_t and use PRIuMAX. */
+		__util_diagf("%s: truncated data line (need %" PRIuMAX " characters, got %" PRIuMAX ")\n",
+			prog, (uintmax_t)needed, (uintmax_t)have);
 		return -1;
 	}
